Split Shape.kt into smaller files
